Pre-process and Source

The msource function runs the macro pre-processor and then executes the program normally.

msource(pth = Sys.path(), file_out = NULL, envir = globalenv(), ...)

Arguments

pth

The path to the R program to source.

file_out

If you want to save or view the generated code from the msource function, supply a full path and file name on this parameter. Default is NULL. When NULL, the function will create a temp file for the generated code.

envir

The environment to be used for program execution. Default is the global environment.

...

Follow-on parameters to the source function. See the source function for additional information.

Value

The results of the source() function, invisibly.

Details

Additional details...

Macro Commands

Below is the complete list of available macro commands:

#%: A macro comment.

#%include '': Inserts code from included file as text into current program.

#%if (): Begins a macro conditional block.

#%elseif (<condition>): Defines a subsequent conditional block.

#%else: Identifies the default behavior in a condition.

#%end: Ends a macro condition.

%sysfunc(): Executes an R expression as part of a macro command.